Added a new MyTicket component for the Internal users so they can also see specifically all of the tickets theyre assigned to instead of having to look through the TicketOverview page.
+ some layout changes (removed old/not neded routes and changes some icons)
